Create a new class for the beta version.
Turns out that you *do* have to change the package name in
AndroidManifest.xml to be the same as the project name in build.xml.
This meant I had to create a new class that imported the entry
point to the program. I'm really glad that package
com.declarative.games.epidemic was only a thin wrapper around
org.libsdl.app.